requirements.txt
mysql-connector-python
faker
tqdm
